Update comment.
The original comment repeated details given in the definition of the 'defs' variable and there was a risk that the comment could become out of date after a code change (e.g. if the Number highlighting group were to be cleared rather than linked to Normal).

The original comment listed only some of the linked highlighting groups.