Abort if running in a terminal that doesn't support 256 colors (or just 2 colors).
By default when running the Windows version of Vim in Windows Terminal the t_Co option is set to 16. (I tried another terminal, cmder - https://cmder.app/ and &t_Co was also 16.) The colors were wrong with &t_Co set to 16, e.g. color 231 (the background color of the Normal highlighting group) was wrong - 231 is white in a 256 color terminal but grey was displayed. The correct colors were used after running ":set t_Co=256". (Windows Terminal v1.17.11461.0, Vim 9.0).

I have no intention of supporting 8 or 16 color terminals. Note that the current color scheme for a 256-color terminal uses a few different shades of gray.

Note regarding new code: we don't check &t_Co if running the GUI. The check that the GUI is running shouldn't be required since Vim v8.2.0554 (which will set &t_Co for the GUI version) but perhaps it's better to do it anyway - A person reading the code might not know that &t_Co will be set for the GUI version.